requirements.txt
psycopg2==2.7.4
requests==2.7.0
python-telegram-bot==10.1.0
gunicorn==19.8.1
feedparser==5.2.1

